requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python
pytest-allure-adaptor
robotframework-python3